Move slugify_with method to text module
